+++ packages/ahg-ric/src/Providers/AhgRicServiceProvider.php
@@ -97,6 +97,7 @@ class AhgRicServiceProvider extends ServiceProvider
                 \AhgRic\Console\Commands\SeedDemo::class,
                 \AhgRic\Console\Commands\ClearGraphCache::class,
                 \AhgRic\Console\Commands\BackfillProductionParticipants::class,
+                \AhgRic\Console\Commands\PurgeOpenWrite::class,
             ]);
         }